Add recent signers check in verifySeal
We don't want a malicious/compromised signer to spam blocks, so a signer is only
allowed to seal 1 block out of every K which is implemented by recent list
logic. However, we miss the check for recent signers in verifying block path.
This commit adds a recent signers check to the verifySeal function.