Correct the parent block headers list passed to snap.apply
The parent block headers list passed to snap.apply is later used in
getValidatorsFromLastCheckpoint to get the last checkpoint block header in case
it is not inserted to local blockchain yet. This list is expected to contain
block headers of previous blocks of current processed block. However, we
mistakenly use the block headers of previous blocks of last snapshot, which
means that list may miss some block headers for last snapshot to current
processed block.