Remove invalid transaction pass transaction from pool
As the verification of transaction pass transaction is done in later phase in
buyGas(), when an invalid transaction pass transaction is identified, it is not
removed from the pool. This commit exposes the removeTx() function from TxPool
and use this function to remove transaction in case it is an invalid transaction
pass transaction.