Add reserved gas for system transactions
Currently, when collecting transactions to commit to a block, we only check the
accumulated gas with block's gas limit. Later, when finalizing the block, with
system transactions included, the block's gas used may exceed the block's gas
limit. As a result, that block is not mined. We want to avoid this scenario by
introducing a new flag to reserve some gas in gas pool for system transactions.