feat: verify the block delay on broadcasted blocks (#193)
This commit follows the approach of Ramanujan in BSC which includes the block
delay in block timestamp for verification on the recipients.

The block timestamp is calculated by the formula
- block timestamp = max(parent timestamp + block duration + backoff time,
  current timestamp)

The block duration is 3s in the current configuration. When the validator is
inturn, the backoff time is 0, otherwise, each out of turn validator has a
different delay.

The delay equals to
- delay = wiggle + (random multiplier * wiggle/2)

The random multiplier is in range [1, len(current validators)]. The random seed
is the block number, the recipient can use the same seed to get the same random
result for verification.