feat: reduce backoff time of not inturn validators (#198)
* feat: reduce backoff time of not inturn validators

Currently, when calculating the backoff time of not inturn validators, we
allocate different delay multipliers to each validator. As a result, the backoff
time ranges from [1, 11] seconds for not inturn validators given the wiggle time
is 1s and number of validators is 21. However, in that turn, some validators
cannot produce the block because they are in recently signed list. So we don't
need to allocate delay multipliers for those validators. Consequently, the
backoff time of not inturn validators shrinks to [1, 6] seconds, this helps to
reduce block time in case some validators are out of service.

* test: add unit test for block's timestamp calculation