test: calculate bloom of each receipt in GenerateChain
In real system, the bloom of each receipt is calculated before
FinalizeAndAssemble block. As a result, we change CreateBloom to MergeBloom when
creating block in NewBlock for optimization. However, the GenerateChain function
which is used mostly for testing does not follow that assumption. This commit
calculates the bloom of each receipt before FinalizeAndAssemble block.