fix: try to load snapshot from disk at forked block before calculating
When trying get snapshot at forked block - 1, we call to contract to get state
at that block which may not be available when chain head is ahead of that block.
When that snapshot is calculated the first time (chain head is at that block, so
the calculation always succeeds), we store that snapshot persistently to the
disk. So, later query to that snapshot, we only need to load it from disk
without re-calculating.