test: fix failed unit tests (#209)
* chore: remove httpdb from Ronin

* test: fix consortium remove invalid recents test

The RemoveOutdatedRecents expects snapshot number which is block number - 1

* test: enable London hardfork in chain config for some tests

* test: rename Geth to Ronin

* test: add test data for wallet import tests

* test: edit the publish event test

This test is outdated since the implementation changes. Update the unit test to
correctly test the current implementation.

* test: update ethclient unit test

We change toBlockNumArg implementation to transform the block number the same
way as RPC: -1 -> latest, -2 -> pending. Update the unit test to reflect the
current implementation.

* test: calculate bloom of each receipt in GenerateChain

In real system, the bloom of each receipt is calculated before
FinalizeAndAssemble block. As a result, we change CreateBloom to MergeBloom when
creating block in NewBlock for optimization. However, the GenerateChain function
which is used mostly for testing does not follow that assumption. This commit
calculates the bloom of each receipt before FinalizeAndAssemble block.