fix: header's hash calculation in verifying double sign proof
We mistakenly calculate hash on the whole header which contains the signature.
This lead to signature malleability issue as the another signature can be
generated on the same block data without knowing the private key. And the check
for same block data is bypassed due to calculating hash on 2 different
signatures. This commit fixes the hash calculation to hash on header excluding
the signature field.