fix: header's hash calculation in verifying double sign proof (#219)
* fix: header's hash calculation in verifying double sign proof

We mistakenly calculate hash on the whole header which contains the signature.
This lead to signature malleability issue as the another signature can be
generated on the same block data without knowing the private key. And the check
for same block data is bypassed due to calculating hash on 2 different
signatures. This commit fixes the hash calculation to hash on header excluding
the signature field.

* test: add ECDSA malleability test for verify double sign proof

We expect our signature verification to be invulnerable to malleability attack.

Co-authored-by: Ta Quang Khanh <khanh.ta@skymavis.com>

* fix: check if the extra data length is greater or equal to signature length

In encodeSigHeader, we assume the extra data length is greater or equal to
signature length, otherwise, we get panic. This commit adds a check in verify to
return earlier.

* fix: check signature in proof is normalized lower-S form

Contract PR: https://github.com/axieinfinity/ronin-dpos-contracts/pull/180 adds
check for already submitted proof but uses the hash over the whole header
including the signature. This is vulnerable to ECDSA malleability attack, check
this in precompiled contract to avoid this issue.

---------

Co-authored-by: Ta Quang Khanh <khanh.ta@skymavis.com>