fix: synchronize concurrent map access in traceInternalsAndAccounts
This PR: https://github.com/axieinfinity/ronin/pull/230 changes the internalTxs
in the response from slice to map which causes panic in CI test because map is
not safe for concurrent use. In order to do this safely, we create a mutex for
synchronization but this may become the bottleneck that defeats the benefit of
multiple concurrent goroutines run in this function.