fix: recently signed check when slashing unavailable validator
The first validator in snap.Recents is shifted out and allowed to seal the block
already. However, when determining if we should slash the validator we mark that
first validator as unable to seal block which is incorrect. This loosens the
slash check rule. This commit fix this bug and create a separate helper function
to check if the validator recently signed a block.