feat: reduce block time of out of turn miner (#252)
This commit takes inspiration from section 5.3 BEP-172: Network Stability
Enhancement On Slash Occur (https://github.com/bnb-chain/BEPs/pull/172)

When a validator is down, its block turn is mined by an out of turn miner. So
when comes the turn of out of turn miner, it may not be able to produce block
due to it is already in recently signed list. This leads to a series of out of
turn blocks which increase the overall block time because currently out of turn
block takes a least 4s. This commit allows out of turn block to take only 3s if
the inturn validator is already in recently signed list.