feat: improve the throttle threshold when syncing
Currently, the new threshold is calculated when fetching the result from result
queue and is applied to the next reserve/fetch result. However, the result
processing may take too much time when data is large, the threshold is not
updated early enough to slow down further reserve/fetch. As a result, the memory
can peak so high. In this commit, we re-calculate threshold when handling
response. In case the result queue is throttled, we discard the response and
queue the fetch task back to queue for later re-fetch. This may increase the
cost in network bandwidth but help to regulate the memory usage.