fix: apply GetBestParentBlock only after Olek hardfork (#271)
Strictly speaking, this logic can apply after some blocks passing the Consortium
v2 hardfork so that when going backward the chain we are still on Consortium v2.
This helps to avoid division by 0 error because of not being able to get
validator list when not passing the Consortium v2. We intend to release this
logic together with Olek hardfork so just check with Olek hardfork here for
symplification.