fix: handle StateByHeader re-entrance when processing Consortium v2 blocks
When processing the block at the end of epoch on Consortium version 2, there is
a call to get block producers of the previous block's state in Finalize.
Because in StateByHeader we use the ephemeral database, the second time we enter
StateByHeader, we use a different in-memory database which does not have the
generated state of previous block. As a result, we need to re-generate that
state which is wasteful. This commit makes the ephemeral database accessible for
the re-entrance to get the previous block's state.