feat: store transaction journal into disk
When processing a block, we execute transactions in the block which results in
the trie changes then later flush the trie changes down to underlying database
layer. This commit reuses the current journal which means to support transaction
revert to store trie changes into disk. Later, when re-processing the block we
can load the journal and apply the trie changes to the trie without applying
transactions.