feat: re-apply journals from nearest snapshot to get missing state
When trying to get a state that is not stored in database, we look up the
nearest snapshot before requested block. Then, re-execute all the stored
journals from snapshot to requested block based on the snapshot state to get the
requested state.

Currently, debug_getModifiedAccountsByNumber/debug_getModifiedAccountsByHash
does not work the same as archive node because the above logic applies to get
state only while these APIs work on triedb directly.