feat: add block journal indexer
When profiling StateByHeader, we see that journal load takes much time because
we need to read the journal from disk every block. This commit tries to reduce
the number of disk read operations by merging the block journals into section. A
journal indexer implementing core.ChainIndexer is created to do that function.