eth/protocols: implement finality vote communication via ronin protocol (#307)
* eth/protocols: implement finality vote communication via ronin protocol

ronin protocol is an extension protocol of eth, a peer is required to support
eth protocol to run ronin protocol. ronin protocol currently only supports
NewVotePacket which contains a list of finality votes.

* eth/protocols: don't broadcast known finality vote to peer

If a vote is broadcasted from a peer or we have broadcasted to that peer
already, we mark this vote is known by that peer and does not broadcast this
vote to that peer again.

* eth/protocols: test finality vote broadcast via ronin protocol