consortium/v2: fix the incorrect snapshot lookup in GetFinalizedBlock
Currently, to check if the justified block number - 1 is justfied, we look at
the snapshot at justfied block number - 1. This is incorrect because at the
snapshot N, the maximum justified block number is N - 1. So, we need to look at
the snapshot at justified block number to check if the justified block number -
1 is justified.