consortium/v2: make SealHash return unchanged hash after Shillin (#332)
As we delay assembling finality vote until Seal function, it means the seal hash
can be updated after the FinalizeAndAssemble. This breaks the worker which uses
seal hash to store and look up seal tasks between FinalizeAndAssemble and Seal
function. Worker uses SealHash function to get the header hash and this function
is used by worker only not by any other parts in the consensus. So in this
commit, we make this function returns a hash of header without the finality
vote. This is not the real hash used for sealing but it remains unchanged after
FinalizeAndAssemble which meet the requirement of the worker.