state_object, trie: update storage trie of an account parallel
We observe the high latency when validating block, which finalises all storage
changes from statedb to trie layer. We want to parallel the trie operations but
parallel means some synchronization methods are need and they cost some
overhead. In this commit, we don't parallel all trie operations, we just
parallel the leaf value node update of the storage trie of an account. This
operation does not change the trie structure, which is simpler to synchronize
comparing to delete or insert a new node.

When updating the leaf node, we need to take care of the hash node resolve, 2
updates cannot resolve the same hash node parallel. We introduce read-write
mutex into the full node and short node to solve this problem. When the child of
these node are hash node, a hash node resolve must happens inside the lock
critical section.