consortium/v2: distribute finality reward for the voters in parent block (#337)
* consortium/v2: distribute finality reward for the voters in parent block

Because the finality votes are assembled after we FinalizeAndAssemble the block
and the finality reward distribution transaction is in FinalizeAndAssemble, we
cannot reward the finality voters in the current block right away. Instead, the
reward distribution is delayed until next block. So in the FinalizeAndAssemble,
we look at the finality voters in the parent header and distribute reward to
them.

* consortium/v2: integrate with the finality tracking contract

This commit integrates with the finality tracking contract
(https://github.com/axieinfinity/ronin-dpos-contracts/pull/279) to record the
finality reward for voters.