state_processor: remove the whitelist deployer check in applyTransaction
We need to update the check in applyTransaction to support the whitelist
deployer version 2. However, since we move the whitelist deployer check to the
later Create operation to cover both CREATE2/CREATE opcode and contract deploy
transaction, this check is not necessary anymore.

Does removing this check create backward incompatibility (cannot sync from
genesis)?
The backward incompatability happens only when the old transactions does not
pass the check and when running with newer Ronin version, it passes because we
remove the check. However, this is not the case here, as the error in
applyTransaction means the transaction is not included into the block.