p2p/server: add DNS resolved candidates to distributed hash table
Currently, DNS resolved candidates of eth protocol are added as candidates for
dialing but not added to DHT. The DHT's seed nodes are still configured via
hardcoded nodes in bootnodes flag. As we want to replace the hardcoded bootnodes
entirely with DNS resolved nodes, we resolve some nodes from the DNS discovery
URL and put them into DHT when setting up the p2p discovery.