tx_pool, state_transition: accept sponsored tx with expired time = 0 (#379)
* tx_pool, state_transition: accept sponsored tx with expired time = 0

The sponsored transaction with expired time = 0 means there is no expired time
for that transaction, so we skip the expired time check for that one.

* api: add payer field to sponsored transaction in RPC response

This commit adds payer field to sponsored transaction in RPC response of APIs
related to transactions to easy lookup without manually recovering from the
returned payer's signature.

* txpool: update the signer of txlist when it differs from txpool's signer

txpool's signer can change when we reach higher block that triggers the new
signer. This commit checks the txpool's and txlist's signer when adding the new
transaction to txlist, update the txlist's signer with txpool's one when they
differ.

* state_transition: nil check gasFeeCap before using in buyGas

In normal case, gasFeeCap is never nil. In transactions other than dynamic fee
transaction, gasFeeCap is set to gasPrice. However, in some test cases,
gasFeeCap can be nil. In these cases, nil check the gasFeeCap and use gasPrice
instead.