eth/filters: reduce database read when handling eth_getLogs
While profiling, we see that when bloom filter hits, 1/3 CPU time in check log
match routine is used to get the corresponding block number from block hash.
This is redundant as this information is available already and can be passed via
function argument. This commit adds block number parameter to GetLogs interface
to pass the available block number information.