consortium: fix not validated snapshot query
The `verifyFinalitySignatures` function is invoked in `VerifyHeaderAndParents`
before the condition
`parent.Number.Uint64() == number-1 && parent.Hash() == header.ParentHash`
is checked. This can result in generating and storing an invalid snapshot
because the block number and block hash provided as inputs for `snapshot`
do not match. If the `if number == c.forkedBlock-1` block is executed,
this can lead to overwriting past snapshot with invalid data.