blockchain_reader: utilize body cache when serving eth_getTransactionReceipt
Currently, when serving eth_getTransactionReceipt, block body is read and
decoded twice in GetTransaction and GetReceipts. This commit makes both
GetTransaction and GetReceipts call to blockchain reader. In blockchain reader,
the handlers of these 2 requests are the same as rawdb handlers but the body
read is cache. As a result, the second body read hits the cache, avoid an
additional database read and body decode.