consortium-v2: cleanup the check of system transaction's sender with coinbase
Currently, the system transaction's sender check looks weird, it still allows
that sender is different from coinbase. However, the sender is checked to be
equal to coinbase in IsSystemMessage already, so there must be no issue here.
This commit fixes that check in ApplyTransaction, we keep this check to make it
more explicit though the check is redundant.