consortium: remove transaction signing logic in getTransactionOpts
The transaction signing logic happens later in ApplyTransaction. The
getTransactionOpts is only used to prepare the transaction data. Furthermore,
the logic in getTransactionOpts does not properly handle block insert path, it
requires that the system transaction signing is executed successfully even in
the block insert path which is not correct. This commit removes the transaction
signing logic in getTransactionOpts to fix the logic when inserting block.